% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{autofit}
\alias{autofit}
\title{Runs a TINC analysis.}
\usage{
autofit(input, cna, VAF_range_tumour = c(0, 0.7),
  cutoff_miscalled_clonal = 0.6, cutoff_lv_assignment = 0.75,
  N = 20000, FAST = FALSE)
}
\arguments{
\item{input}{A`dataframe` of the iput mutations. Must be in a certain format,
see the vignette for more information.}

\item{cna}{Copy Number data in the format of package \code{CNAqc}.}

\item{VAF_range_tumour}{A range `[x, y]` so that only mutations
with VAF in that range are actually used to determine the TIN/ TIT
levels of the input.}

\item{cutoff_miscalled_clonal}{An upper bound on the VAF of a
cluster in the tumour data. Clusters above this value will be
considered miscalled clonal clusters (e.g., due to LOH etc.).}

\item{cutoff_lv_assignment}{Consider only latent variables with
responsibilities above this cutoff.}

\item{N}{If there are more than `N` mutations in VAF range
`VAF_range_tumour`, a random subset of size `N` is retained.}

\item{FAST}{If `TRUE`, it runs the analysis with reduced sampling
power and accuracy. Use this to obtain a result for preliminary
inspection of your data, and then run `autofit` with this parameter
set to `FALSE`.}
}
\value{
An S3 object that contains the results of this analysis.
}
\description{
This function is a wrapper to run the main analysis of TINC.

The steps are as follows:
\itemize{
\item 1) Input data is loaded from a `file`, or from a `dataframe`.
\item 2) Clonal mutations are estimated for the tumour, together with
the tumour purity (Tumour in Tumour).
\item 3) From putative clonal mutations of the tumour, the Tumour in Normal
contamination level is estimated.
}
An S3 object is returned that contains the results of the analysis.
}
\examples{
# Random
rt = TINC::random_TIN()
x = TINC::autofit(input = rt$data, cna = rt$cna, FAST = TRUE)
print(x)
}
