% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_CHIP_variants.R
\name{check_CHIP_variants}
\alias{check_CHIP_variants}
\title{Check for CHIP variants in normal and tumour at high VAF}
\usage{
check_CHIP_variants(
  normal_bam = NULL,
  tumour_bam = NULL,
  reference = "hg38",
  min_vaf = 0.05,
  only_snvs = TRUE,
  add_chr = TRUE
)
}
\arguments{
\item{normal_bam}{path of bam file for healthy sample}

\item{tumour_bam}{path of bam file for tumour sample}

\item{reference}{reference genome, one of hg38 or hg19}

\item{min_vaf}{minimum required VAF to flag the mutation}

\item{only_snvs}{use only SVNs and filter indels}

\item{add_chr}{if chromosomes on BAM files have chr set it to TRUE

The function takes a tumour and a normal bam files and scans for known CHIP mutations listed in \href{https://www.nature.com/articles/s41586-020-2819-2}{Bick et al.}. The function gives back a data frame with mutation having VAF higher than \code{min_vaf}.
A bit of caution has to be taken when interpreting indels results, as from the pileup we cannot assign a lenght to the indel and the frameshit nature of the alteration. We suggest doing a proper small variant calling to confirm the results.}
}
\value{
Two dataframes for each of tumour and nomal with coloumns: chr, pos, alt, ref, VAF, DP and NV (chromosome, position, alternative allele, reference allele, variant allelic frequency, depth , number of variant reads)
}
\description{
Check for CHIP variants in normal and tumour at high VAF
}
