% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_example}
\alias{fit_example}
\title{Example TINC fit}
\format{
Fit from TINC.
}
\usage{
data(fit_example)
}
\description{
Example TINC fit
}
\examples{
data(fit_example)
fit_example

}
\references{
Simulated data.
}
\keyword{datasets}
