% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_files.R
\name{load_input}
\alias{load_input}
\title{Load data for TIN.}
\usage{
load_input(x, verbose = FALSE, VAF_range_tumour = c(0, 0.7),
  N = 20000)
}
\arguments{
\item{x}{Either a filename, or a dataframe.}

\item{verbose}{If `TRUE`, more output is print to screen.}

\item{VAF_range_tumour}{A 2D vector that determines a
VAF range used to decide which mutation we do analyse from the
input cancer sample.}

\item{N}{If there are more than `N` mutations in VAF range
`VAF_range_tumour`, a random subset of size `N` is retained.}
}
\value{
A list that contains the following informations:
tumour calls, normal calls and joint calls, plus the input
VAF range and file name.
}
\description{
The function can support both a file argument and a dataframe.
In the case a file is given, no column or row names should be
used. If the input is a dataframe or tibble, column names are
required. See the package for a detailed description of the
input format for TIN analysis.

This function, after loading data, annotates which mutation
has VAF  within a custom range (default [0; 0.7]); these are
the only mutations that are used for the TIN analysis.
}
\examples{
# Generating a random TIN input
load_input(random_TIN())
}
