% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_TINC_input.R
\name{load_TINC_input}
\alias{load_TINC_input}
\title{Load TINC input data.}
\usage{
load_TINC_input(x, cna, VAF_range_tumour = c(0, 0.7), N = 20000)
}
\arguments{
\item{x}{A dataframe or tibble with input mutation data, reporting `chr`, `from`, `to`, `ref` and `alt`,
plus `n_ref_count` and `n_alt_count`, and `t_ref_count` and `t_tot_count`.}

\item{cna}{Copy Number data in the format of package \code{CNAqc}, providing `chr`, `from`, `to`,
`Major` and `minor`.}

\item{VAF_range_tumour}{VAF range used to filter
mutations from the tumour sample.}

\item{N}{If there are more than `N` mutations in VAF range
`VAF_range_tumour`, a random subset of size `N` is retained.}
}
\value{
A tibble with the loaded data.
}
\description{
The function loads input mutations and optional copy number data for TINC.
Input formats are reported at the package website.

After loading data, mutation with VAF outside a range (default [0; 0.7])
are removed from analysis. Similarly, mutations are down-sampled if in excess of
some  threshold (default `N=20000`).
}
\examples{
# Generating a random TIN input
rt = random_TIN()
load_TINC_input(x = rt$data, cna = rt$cna)
}
