% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{load_VCF_Canvas}
\alias{load_VCF_Canvas}
\title{Canvas VCF parsing function}
\usage{
load_VCF_Canvas(file)
}
\arguments{
\item{file}{VCF filename.}
}
\value{
A named list with the calls and tumour purity.
}
\description{
Parse a VCF file from Canvas, which stores a
list of segments with absolute CNA events, and tumour purity.

The obtained data, together with mutation calls (loaded elsewhere)
can be used to call function \code{autofit}.
}
\examples{
# not run


}
