% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{load_VCF_CNA_DRAGEN}
\alias{load_VCF_CNA_DRAGEN}
\title{DRAGEN VCF CNA parsing function}
\usage{
load_VCF_CNA_DRAGEN(file)
}
\arguments{
\item{file}{is DRAGEN CNV vcf file path}
}
\value{
PASS copy number alteration values in CNAqc format, stored in the `cna`
slot, and estimated tumour purity in the `purity` slot.
}
\description{
Parse a VCF CNA file from DRAGEN, which has information on both
}
\examples{
# not run
\dontrun{
 load_VCF_CNA_DRAGEN("Myfile.vcf")
}
}
