% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{load_VCF_smallVar_MUTECT2}
\alias{load_VCF_smallVar_MUTECT2}
\title{Mutect2 VCF small variants parsing function}
\usage{
load_VCF_smallVar_MUTECT2(file)
}
\arguments{
\item{file}{is Mutect2 small variants vcf file path}
}
\value{
Allele depths for normal and tumour
for all PASS autosome SNVs
}
\description{
Parse a small variants VCF file from Mutect2 with information about SNVs and indels
}
\examples{
# not run
\dontrun{
 load_VCF_smallVar_MUTECT2("Myfile.vcf")
}
}
