% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{load_VCF_Strelka}
\alias{load_VCF_Strelka}
\title{Strelka 2 VCF parsing function}
\usage{
load_VCF_Strelka(file)
}
\arguments{
\item{file}{is Strelka 2 vcf file path}
}
\value{
Allele depths for normal and tumour
for all PASS autosome SNVs
}
\description{
Parse a VCF file from Srelka 2, which stores for
PASS SNVs the read depth of the alleles in the
nornmal and tumour
}
\examples{
# not run
}
