% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{plot.tin_obj}
\alias{plot.tin_obj}
\title{Plot a TINC analysis.}
\usage{
plot.tin_obj(x, ...)
}
\arguments{
\item{x}{A TINC analysis computed with \code{autofit}.}

\item{...}{Extra S3 parameters.}
}
\value{
A TINC analysis plot (a `ggplot` figure with multiple panels).
}
\description{
Plot a TINC analysis.
}
\examples{
plot(autofit(random_TIN(), FAST = TRUE))
}
