% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_full_page_report.R
\name{plot_full_page_report}
\alias{plot_full_page_report}
\title{Plot a full page report of the TINC analysis}
\usage{
plot_full_page_report(x)
}
\arguments{
\item{x}{}
}
\description{
The report is a multi-panel figure that contains the following
information.
}
