% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simple_report.R
\name{plot_simple_report}
\alias{plot_simple_report}
\title{Plot a strip short report of the TINC analysis.}
\usage{
plot_simple_report(x)
}
\arguments{
\item{x}{A TINC analysis computed with \code{autofit}.}
}
\value{
A multi-panel \code{ggplot} figure.
}
\description{
The report is a multi-panel figure that contains the following
information. TODO fill in.
}
\examples{
plot_simple_report(autofit(random_TIN(), FAST = TRUE))
}
