% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{pullAD}
\alias{pullAD}
\title{Allele depth pull function}
\usage{
pullAD(CHROM, POS, REF, ALT, FILTER, FORMAT, normal, tumour)
}
\arguments{
\item{as}{columns in vcf}
}
\value{
Allele depths for normal and tumour
}
\description{
Pulls out normal and tumour allele depths from
Strelka 2 vcf columns
}
\examples{
# not run
}
