% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_string.R
\name{to_string}
\alias{to_string}
\title{Create a data frame representation of a TINC fit}
\usage{
to_string(x)
}
\arguments{
\item{x}{A TINC fit.}
}
\value{
A dataframe.
}
\description{
Create a data frame representation of a TINC fit
}
