% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_TINC_input.R
\name{load_TINC_input}
\alias{load_TINC_input}
\title{Load TINC input data.}
\usage{
load_TINC_input(x, cna, VAF_range_tumour = c(0, 0.7), N = 20000)
}
\arguments{
\item{x}{A dataframe or tibble.}

\item{cna}{Copy Number data in the format of package \code{CNAqc}.}

\item{VAF_range_tumour}{2D vector for a VAF range used to filter
mutations from the tumour sample.}

\item{N}{If there are more than `N` mutations in VAF range
`VAF_range_tumour`, a random subset of size `N` is retained.}
}
\value{
A tibble with the loaded data.
}
\description{
The function loads a dataframe or tibble, with a set of required
column names. See the package for a detailed description of the
input format.

After loading data, mutation with VAF outside a range (default [0; 0.7])
are flagged and removed from downnstream analysis.
}
\examples{
# Generating a random TIN input
load_TINC_input(random_TIN())
}
