% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_full_page_report.R
\name{plot_full_page_report}
\alias{plot_full_page_report}
\title{Plot a full page detailed report of the TINC analysis}
\usage{
plot_full_page_report(x)
}
\arguments{
\item{x}{A TINC analysis computed with \code{autofit}.}
}
\value{
A multi-panel \code{ggplot} figure.
}
\description{
The report is a multi-panel figure that contains the following
information. TODO fill in.
}
\examples{
rt = random_TIN()
plot_full_page_report(autofit(input = rt$data, cna = rt$cna, FAST = TRUE))
}
