% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{print.tin_obj}
\alias{print.tin_obj}
\title{Print a TINC object to screen}
\usage{
\method{print}{tin_obj}(x, ...)
}
\arguments{
\item{x}{A TINC analysis computed with \code{autofit}.}

\item{...}{Extra S3 parameters}
}
\value{

}
\description{
Print a TINC object to screen
}
\examples{
# Automatic call
autofit(random_TIN(), FAST = TRUE)
}
