% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{pullAD}
\alias{pullAD}
\title{Allele depth pull function}
\usage{
pullAD(CHROM, POS, REF, ALT, FILTER, FORMAT, normal, tumour)
}
\arguments{
\item{CHROM}{as CHROM in vcf}

\item{POS}{as POS in vcf}

\item{REF}{as REF in vcf}

\item{ALT}{as ALT in vcf}

\item{FILTER}{as FILTER in vcf}

\item{FORMAT}{as FORMAT in vcf}

\item{normal}{as normal in vcf}

\item{tumour}{as tumour in vcf}
}
\value{
Allele depths for normal and tumour
}
\description{
Pulls out normal and tumour allele depths from
Strelka 2 vcf columns
}
\examples{
# not run
}
