% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_TIN.R
\name{random_TIN}
\alias{random_TIN}
\title{Title}
\usage{
random_TIN(N = 1000, TIN = 0.05, TIT = 1, normal_coverage = 30,
  tumour_coverage = 120)
}
\arguments{
\item{tumour_coverage}{}
}
\description{
Title
}
\examples{
random_TIN()
}
